+++ ext/model.h
@@ -21,13 +21,14 @@ private:
 public:
     Model(const std::string filename, bool diffuse_texture = false, bool normal_map = false,
           bool specular_texture = false);
-    int nverts() const;
-    int nfaces() const;
-    vec3 normal(const int iface, const int nthvert) const;  // per triangle corner normal vertex
-    vec3 normal(const vec2 &uv) const;  // fetch the normal vector from the normal map texture
-    vec3 vert(const int i) const;
-    vec3 vert(const int iface, const int nthvert) const;
-    vec2 uv(const int iface, const int nthvert) const;
+    size_t nverts() const;
+    size_t nfaces() const;
+    vec3 normal(const size_t iface,
+                const size_t nthvert) const;  // per triangle corner normal vertex
+    vec3 normal(const vec2 &uv) const;        // fetch the normal vector from the normal map texture
+    vec3 vert(const size_t i) const;
+    vec3 vert(const size_t iface, const size_t nthvert) const;
+    vec2 uv(const size_t iface, const size_t nthvert) const;
     TGAColor diffuse(const vec2 &uv) const;
     double specular(const vec2 &uv) const;
 };
